Updates to fix a few issues and add functionality
The SpeechCenter had an issue with the way I was passing the text string to the command line for the system to speak, which resulted in it having issues with apostrophes and with using SSML (Speech Synthesis Markup Language).  That being fixed with this update, SSML has now been enabled, giving greater control over the speech output, allowing for subtle (or not-so-subtle) emphasis additions, pitch changes, rate-of-speech adjustments and even highly precise pronunciation adjustments.  In addition to teh code changes in SpeechCenter, a summary of common SSML examples are given in teh comments at the bottom pf the code.